Extract favorites helpers into submodule
Convert backend/routes/favorites.py to a package. The 13 module-level
helpers (serializers, project/tag loaders, agreement resolvers, no-store
wrapper) move to backend/routes/favorites/helpers.py. FavoritesDeps moves
to a tiny backend/routes/favorites/deps.py so helpers can import the type
without circularity.

__init__.py drops from 1135 to 922 LOC. Public surface (FavoritesDeps,
register_favorites_routes) is unchanged.
